requirements.txt
streamlit
openai
numpy
scikit-learn
pdfminer.six
python-docx
tqdm
langchain
python-dotenv
pandas
sentence-transformers
streamlit-lottie
